requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-allauth==65.4.1
django-ckeditor-5==0.2.17
django-environ==0.12.0
django-js-asset==3.1.2
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-taggit==6.1.0
django-widget-tweaks==1.5.0
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.1
sentry-sdk==2.22.0
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
